requirements.txt
flask
nltk
gunicorn
Flask-Failsafe
psycopg2
Flask-SQLAlchemy
Flask-Migrate
alembic
celery
factory-boy
Flask-Migrate
Flask-RESTful
Flask-Script
Jinja2
six
SQLAlchemy-Utils
Werkzeug
psycopg2
pytest
